use post fetch for sid cookie
